requirements.txt
geopandas
shapely
rasterio
numpy
pandas
pyproj
pyogrio
pyarrow
pillow
requests
scipy
folium
matplotlib
contextily
